Create an xunit results file even if no tests are run.
This PR ensures that when `--xunit-output` is passed to `swift test --parallel`, the output XML file is generated even if no tests are run.

Failure modes outside of test failures (i.e. errors thrown during the build process or when trying to run the test process) will still prevent the generation of this file.

Resolves #7065.